Add Watch Tracking
The player now sends POST requests to /user/tracking/[name] containing the
current show, season and episode, as well as the progress through the episode
as a percentage, on video TimeUpdates.

The server handles these requests by writing the information to the user's
watched.json (/static/users/[name]/watched.json).

This information is not used by the client yet.